# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Student AI Toolkit API",
        "description": "API for the Student AI Toolkit",
        "version": "1.0.0",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
        "/image/upload": {
            "post": {
                "tags": [
                    "images"
                ],
                "summary": "Upload an image",
                "description": "Upload an image",
                "operationId": "uploadImage",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "image",
                        "in": "formData",
                        "description": "The image to upload",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "ID of the user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image uploaded"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/image/{userId}": {
            "get": {
                "tags": [
                    "images"
                ],
                "summary": "Get user image",
                "description": "Get user image",
                "operationId": "getUserImage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User image found"
                    },
                    "404": {
                        "description": "User image not found"
                    }
                }
            }
        },
...
        "/readPdf": {
            "post": {
              "tags": [
                "file"
              ],
              "summary": "Read a PDF file",
              "description": "Read a PDF file uploaded as form data",
              "operationId": "readPdf",
              "consumes": [
                "multipart/form-data"
              ],
              "produces": [
                "application/json"
              ],
              "parameters": [
                {
                  "in": "formData",
                  "name": "title",
                  "description": "Title of the PDF file",
                  "required": true,
                  "type": "string"
                },
                {
                  "in": "formData",
                  "name": "pdf",
                  "description": "PDF file to be uploaded",
                  "required": true,
                  "type": "file"
                }
              ],
              "responses": {
                "200": {
                  "description": "PDF file read successfully",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      },
                      "file": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          },
                          "text": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
        },
        "/files": {
            "get": {
              "tags": [
                "files"
              ],
              "summary": "Get all files",
              "description": "Get a list of all files",
              "operationId": "getAllFiles",
              "produces": [
                "application/json"
              ],
              "responses": {
                "200": {
                  "description": "List of files",
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/File"
                    }
                  }
                },
                "500": {
                  "description": "Internal server error",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
        },
        "/files/{id}": {
            "get": {
              "tags": [
                "files"
              ],
              "summary": "Get file by ID",
              "description": "Get a file by its ID",
              "operationId": "getFileById",
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "description": "ID of the file",
                  "required": true,
                  "type": "string"
                }
              ],
              "produces": [
                "application/json"
              ],
              "responses": {
                "200": {
                  "description": "File found",
                  "schema": {
                    "$ref": "#/definitions/File"
                  }
                },
                "404": {
                  "description": "File not found"
                },
                "500": {
                  "description": "Internal server error",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
        },
        "/file/{id}": {
            "delete": {
                "tags": [
                  "files"
                ],
                "summary": "Delete file by ID",
                "description": "Delete a file by its ID",
                "operationId": "deleteFileById",
                "parameters": [
                  {
                    "name": "id",
                    "in": "path",
                    "description": "ID of the file",
                    "required": true,
                    "type": "string"
                  }
                ],
                "responses": {
                  "200": {
                    "description": "File deleted successfully"
                  },
                  "404": {
                    "description": "File not found"
                  },
                  "500": {
                    "description": "Internal server error",
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
        },          
...
    },
...
}